Add snappy compression option to Barman Cloud #344
---

The implementation of decompressing streaming IO is the same for AWS and
Azure aside from the details of how they actually read the compressed
bytes from the stream.

We therefore move them into a mixin which can be inherited from when
subclassing the Streaming IO response for each cloud provider. The
provider-specific logic to read the compressed bytes from the stream is
then added in the subclass.

---

Increase the buffer sizes to 65535. Lower values yield either poorer compression
ratios or slower decompression times. We use 65535 because the python-snappy comments
suggest it is suitable for most scenarios and anecdotal testing confirms this.