Add Snappy compression option to Barman Cloud #344
Adds the `--snappy` option to barman-cloud-backup and
barman-cloud-wal-archive which causes Barman Cloud to apply the Snappy
compression algorithm to the data uploaded to the cloud provider. This
yields faster performance but poorer compression ratios than gzip and
bzip2.

As with bzip2 and gzip compression, The barman-cloud-restore and
barman-cloud-wal-archive will automatically detect when Snappy
compression has been applied and decompress using Snappy while
downloading.

Because Snappy compression requires external dependencies but may not be
useful to every Barman user, the dependency is handled as follows:

1. The python-snappy library is added to `extras_require` in setup.py
   rather than being made a core requirement.
2. Any code which needs to interact with the snappy module is moved into
   a cloud_compression module where snappy is lazily imported *only* if
   snappy compression is required.

This means that users will need to install python-snappy themselves if
they want to use the `--snappy` option.

---

The implementation of decompressing streaming IO is the same for AWS and
Azure aside from the details of how they actually read the compressed
bytes from the stream.

We therefore move them into a mixin which can be inherited from when
subclassing the Streaming IO response for each cloud provider. The
provider-specific logic to read the compressed bytes from the stream is
then added in the subclass.

---

Increase the buffer sizes to 65535. Lower values yield either poorer compression
ratios or slower decompression times. We use 65535 because the python-snappy comments
suggest it is suitable for most scenarios and anecdotal testing confirms this.